change priority to int

+++ src/Microsoft.AspNetCore.Mvc.Modules.Abstractions/LocationExpander/IViewLocationExpanderProvider.cs
@@ -4,6 +4,6 @@ namespace Microsoft.AspNetCore.Mvc.Modules.LocationExpander
 {
     public interface IViewLocationExpanderProvider : IViewLocationExpander
     {
-        double Priority { get; }
+        int Priority { get; }
     }
 }
